+++ junixsocket-common/src/main/java/org/newsclub/net/unix/AFSocketChannel.java
@@ -131,7 +131,7 @@ public abstract class AFSocketChannel<A extends AFSocketAddress> extends SocketC
     if (optionId == null) {
       throw new UnsupportedOperationException("unsupported option");
     } else {
-      return (T) afSocket.getAFImpl().getOption(optionId.intValue());
+      return (T) afSocket.getAFImpl().getOption(optionId);
     }
   }
 
@@ -145,7 +145,7 @@ public abstract class AFSocketChannel<A extends AFSocketAddress> extends SocketC
     if (optionId == null) {
       throw new UnsupportedOperationException("unsupported option");
     } else {
-      afSocket.getAFImpl().setOption(optionId.intValue(), value);
+      afSocket.getAFImpl().setOption(optionId, value);
     }
     return this;
   }
